feat(alert): 탭, 카테고리 사용되는 state제거 및 로직 간소화/분리
