+++ src/app/(route)/alert/_constants/MOCK_ALERT_ITEMS.ts
@@ -2,6 +2,7 @@ export const MOCK_ALERT_ITEMS = [
   {
     id: 1,
     category: "전자기기",
+    categoryKey: "categoryKeyword" as const,
     title: ` 새 게시글이 등록되었어요.`,
     body: "시청역 아이폰 14 보관중입니다.",
     createdAt: "2025-11-25T11:15:04.559Z",
@@ -12,6 +13,7 @@ export const MOCK_ALERT_ITEMS = [
   {
     id: 2,
     category: "확인하지 않은 채팅",
+    categoryKey: "chat" as const,
     title: "이 있어요.",
     body: "혹시 OO동 OO빌딩 화장실에서 핸드폰 잃어버리셨나요 그럼 혹시",
     createdAt: "2025-11-25T11:20:04.559Z",
@@ -22,6 +24,7 @@ export const MOCK_ALERT_ITEMS = [
   {
     id: 3,
     category: "새로운 채팅",
+    categoryKey: "chat" as const,
     title: "이 도착했어요.",
     body: "혹시 OO동 OO빌딩 화장실에서 핸드폰 잃어버리셨나요 그럼 혹시",
     createdAt: "2025-11-20T11:15:04.559Z",
@@ -32,6 +35,7 @@ export const MOCK_ALERT_ITEMS = [
   {
     id: 4,
     category: "즐겨찾기한 게시글의 상태",
+    categoryKey: "categoryKeyword" as const,
     title: "가 변경되었어요.",
     body: "노원구에서 검정색 카드 지갑 잃어버렸어요",
     createdAt: "2025-11-18T11:15:04.559Z",
@@ -42,6 +46,7 @@ export const MOCK_ALERT_ITEMS = [
   {
     id: 5,
     category: "새로운 댓글",
+    categoryKey: "comment" as const,
     title: "이 달렸어요.",
     body: "서울시 노원구 OO동 건물 화장실에서 핸드폰을 잃어버렸어요. 핸드폰은",
     createdAt: "2025-11-19T11:15:04.559Z",
